+++ HPXMLtoOpenStudio/measure.rb
@@ -62,6 +62,12 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     arg.setDefaultValue('results_annual')
     args << arg
 
+    arg = OpenStudio::Measure::OSArgument::makeStringArgument('electric_panel_output_file_name', false)
+    arg.setDisplayName('Electric Panel Output File Name')
+    arg.setDescription("The name of the file w/ electric panel outputs. If not provided, defaults to 'results_panel.csv' (or '.json' or '.msgpack').")
+    arg.setDefaultValue('results_panel')
+    args << arg
+
     arg = OpenStudio::Measure::OSArgument::makeStringArgument('design_load_details_output_file_name', false)
     arg.setDisplayName('Design Load Details Output File Name')
     arg.setDescription("The name of the file w/ additional HVAC design load details. If not provided, defaults to 'results_design_load_details.csv' (or '.json' or '.msgpack').")
@@ -163,6 +169,17 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
       Outputs.append_sizing_results(hpxml.buildings, annual_results_out)
       Outputs.write_results_out_to_file(annual_results_out, args[:output_format], args[:annual_output_file_path])
 
+      # Write electric panel output file
+      # Currently, we only write this if:
+      # (1) at least one load calculation type is specified
+      # (2) an electric panel is specified
+      if (not hpxml.header.service_feeders_load_calculation_types.empty?) &&
+         (hpxml.buildings.map { |hpxml_bldg| hpxml_bldg.electric_panels.size }.sum > 0)
+        electric_panel_results_out = Outputs.append_panel_results(hpxml.header, hpxml.buildings)
+
+        Outputs.write_results_out_to_file(electric_panel_results_out, args[:output_format], args[:electric_panel_output_file_path])
+      end
+
       # Write design load details output file
       HVACSizing.write_detailed_output(design_loads_results_out, args[:output_format], args[:design_load_details_output_file_path])
     rescue Exception => e
@@ -193,6 +210,11 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
       args[:annual_output_file_name] = "#{args[:annual_output_file_name]}.#{args[:output_format]}"
     end
     args[:annual_output_file_path] = File.join(args[:output_dir], args[:annual_output_file_name])
+
+    if File.extname(args[:electric_panel_output_file_name]).length == 0
+      args[:electric_panel_output_file_name] = "#{args[:electric_panel_output_file_name]}.#{args[:output_format]}"
+    end
+    args[:electric_panel_output_file_path] = File.join(args[:output_dir], args[:electric_panel_output_file_name])
 
     if File.extname(args[:design_load_details_output_file_name]).length == 0
       args[:design_load_details_output_file_name] = "#{args[:design_load_details_output_file_name]}.#{args[:output_format]}"