+++ HPXMLtoOpenStudio/tests/test_electric_panel.rb
@@ -664,6 +664,92 @@ class HPXMLtoOpenStudioElectricPanelTest < Minitest::Test
     end
   end
 
+  def test_electric_panel_output_file
+    args_hash = { 'output_format' => 'json',
+                  'hpxml_path' => File.absolute_path(File.join(@sample_files_path, 'base-detailed-electric-panel.xml')) }
+    _model, hpxml, _hpxml_bldg = _test_measure(args_hash)
+    electric_panel_path = File.absolute_path(File.join(File.dirname(__FILE__), 'results_panel.json'))
+    json = JSON.parse(File.read(electric_panel_path))
+
+    assert_equal(16, json['Electric Panel Breaker Spaces']['Total Count'])
+    assert_equal(11, json['Electric Panel Breaker Spaces']['Occupied Count'])
+    assert_equal(16 - 11, json['Electric Panel Breaker Spaces']['Headroom Count'])
+    assert_equal(9444.2, json['Electric Panel Load']['2023 Existing Dwelling Load-Based: Total Load (W)'])
+    assert_equal(39.4, json['Electric Panel Load']['2023 Existing Dwelling Load-Based: Total Capacity (A)'])
+    assert_in_epsilon(100.0 - 39.4, json['Electric Panel Load']['2023 Existing Dwelling Load-Based: Headroom Capacity (A)'], 0.01)
+    assert_equal(5625.0, json['Electric Panel Load']['2023 Existing Dwelling Meter-Based: Total Load (W)'])
+    assert_equal(23.4, json['Electric Panel Load']['2023 Existing Dwelling Meter-Based: Total Capacity (A)'])
+    assert_in_epsilon(100.0 - 23.4, json['Electric Panel Load']['2023 Existing Dwelling Meter-Based: Headroom Capacity (A)'], 0.01)
+
+    # Upgrade
+    hpxml_bldg = hpxml.buildings[0]
+    electric_panel = hpxml_bldg.electric_panels[0]
+    electric_panel.rated_total_spaces = 16
+    branch_circuits = electric_panel.branch_circuits
+    service_feeders = electric_panel.service_feeders
+    sf = service_feeders.find { |sf| sf.type == HPXML::ElectricPanelLoadTypeHeating }
+    sf.power = 16942
+    sf.is_new_load = true
+    branch_circuits.add(id: "BranchCircuit#{branch_circuits.size + 1}",
+                        voltage: HPXML::ElectricPanelVoltage240,
+                        occupied_spaces: 5,
+                        component_idrefs: [hpxml_bldg.heating_systems[0].id])
+    sf = service_feeders.find { |sf| sf.type == HPXML::ElectricPanelLoadTypeCooling }
+    sf.power = 16942
+    sf.is_new_load = true
+    branch_circuits.add(id: "BranchCircuit#{branch_circuits.size + 1}",
+                        voltage: HPXML::ElectricPanelVoltage240,
+                        occupied_spaces: 0,
+                        component_idrefs: [hpxml_bldg.cooling_systems[0].id])
+    service_feeders.add(type: HPXML::ElectricPanelLoadTypeWaterHeater,
+                        power: 4500,
+                        is_new_load: true,
+                        component_idrefs: [hpxml_bldg.water_heating_systems[0].id])
+    branch_circuits.add(id: "BranchCircuit#{branch_circuits.size + 1}",
+                        voltage: HPXML::ElectricPanelVoltage240,
+                        occupied_spaces: 2,
+                        component_idrefs: [hpxml_bldg.water_heating_systems[0].id])
+    service_feeders.add(type: HPXML::ElectricPanelLoadTypeClothesDryer,
+                        power: 5760,
+                        is_new_load: true,
+                        component_idrefs: [hpxml_bldg.clothes_dryers[0].id])
+    branch_circuits.add(id: "BranchCircuit#{branch_circuits.size + 1}",
+                        voltage: HPXML::ElectricPanelVoltage240,
+                        occupied_spaces: 2,
+                        component_idrefs: [hpxml_bldg.clothes_dryers[0].id])
+    service_feeders.add(type: HPXML::ElectricPanelLoadTypeRangeOven,
+                        power: 12000,
+                        is_new_load: true,
+                        component_idrefs: [hpxml_bldg.cooking_ranges[0].id])
+    branch_circuits.add(id: "BranchCircuit#{branch_circuits.size + 1}",
+                        voltage: HPXML::ElectricPanelVoltage240,
+                        occupied_spaces: 2,
+                        component_idrefs: [hpxml_bldg.cooking_ranges[0].id])
+    hpxml_bldg.plug_loads.add(id: "PlugLoad#{hpxml_bldg.plug_loads.size + 1}",
+                              plug_load_type: HPXML::PlugLoadTypeElectricVehicleCharging)
+    service_feeders.add(type: HPXML::ElectricPanelLoadTypeElectricVehicleCharging,
+                        power: 1650,
+                        is_new_load: true,
+                        component_idrefs: [hpxml_bldg.plug_loads[-1].id])
+    XMLHelper.write_file(hpxml.to_doc(), @tmp_hpxml_path)
+
+    args_hash['hpxml_path'] = @tmp_hpxml_path
+    args_hash['skip_validation'] = true
+    _model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
+    electric_panel_path = File.absolute_path(File.join(File.dirname(__FILE__), 'results_panel.json'))
+    json = JSON.parse(File.read(electric_panel_path))
+
+    assert_equal(16, json['Electric Panel Breaker Spaces']['Total Count'])
+    assert_equal(11 + 12, json['Electric Panel Breaker Spaces']['Occupied Count'])
+    assert_equal(16 - 23, json['Electric Panel Breaker Spaces']['Headroom Count'])
+    assert_equal(34827.2, json['Electric Panel Load']['2023 Existing Dwelling Load-Based: Total Load (W)'])
+    assert_equal(145.1, json['Electric Panel Load']['2023 Existing Dwelling Load-Based: Total Capacity (A)'])
+    assert_in_epsilon(100.0 - 145.1, json['Electric Panel Load']['2023 Existing Dwelling Load-Based: Headroom Capacity (A)'], 0.01)
+    assert_equal(46477.0, json['Electric Panel Load']['2023 Existing Dwelling Meter-Based: Total Load (W)'])
+    assert_equal(193.7, json['Electric Panel Load']['2023 Existing Dwelling Meter-Based: Total Capacity (A)'])
+    assert_in_epsilon(100.0 - 193.7, json['Electric Panel Load']['2023 Existing Dwelling Meter-Based: Headroom Capacity (A)'], 0.01)
+  end
+
   private
 
   def _test_service_feeder_power(hpxml_bldg, type, power)